+++ src/main/java/org/jpokemon/api/evolutions/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Provides classes related to a Pokémon's potential `evolution` path.
  */
-package org.jpokemon.api.pokemon.evolution;
+package org.jpokemon.api.evolutions;
